18-6

+++ src/pages/PrivateTravel/PrivateTrip.jsx
@@ -14,6 +14,7 @@ const PrivateTrip = () => {
   const [acceptedUsers, setAcceptedUser] = useState([]);
   const [selectedUser, setSelectedUser] = useState(null);
   const [isPopupVisible, setIsPopupVisible] = useState(false);
+  const [deleteORaccept, setdeleteORaccept] = useState(false);
 
   const [users, setUsers] = useState([]); 
   const token = localStorage.getItem('token');
@@ -68,26 +69,51 @@ const PrivateTrip = () => {
   };
 
   const handleDeleteSelected = () => {
+    setdeleteORaccept(false);
+    setIsPopupVisible(true);
+  };
+
+  const handleAcceptedSelected = () => {
+    setdeleteORaccept(true);
     setIsPopupVisible(true);
   };
 
   const confirmDelete = async () => {
     setIsPopupVisible(false);
-    try {
-      const idsToDelete = selectedUser ? [selectedUser.id] : selectedUsers;
-      const response = await axios.delete('https://jawak-wa-tareekak.onrender.com/jawak-wa-tareekak/manager/travels/private', {
-        data: { ids: idsToDelete },
-        headers: {
-          'Authorization': `Bearer ${token}`,
-          'Content-Type': 'application/json'
-        }
-      });
-      fetchUsers();
-    } catch (error) {
-      console.error('Error deleting extra', error);
+    if(!deleteORaccept) {
+      try {
+        const idsToDelete = selectedUser ? [selectedUser.id] : selectedUsers;
+        const response = await axios.delete('https://jawak-wa-tareekak.onrender.com/jawak-wa-tareekak/manager/travels/private', {
+          data: { ids: idsToDelete },
+          headers: {
+            'Authorization': `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        });
+        fetchUsers();
+      } catch (error) {
+        console.error('Error deleting extra', error);
+      }
+      setSelectedUsers([]);
+      setSelectedUser(null);
     }
-    setSelectedUsers([]);
-    setSelectedUser(null);
+
+    if(deleteORaccept) {
+      try {
+        const idsToAccept = acceptedUsers;
+        const response = await axios.put('https://jawak-wa-tareekak.onrender.com/jawak-wa-tareekak/manager/travels/update-many', {
+          data: { ids: idsToAccept, status: 'accepted' },
+          headers: {
+            'Authorization': `Bearer ${token}`,
+            'Content-Type': 'application/json'
+          }
+        });
+        fetchUsers();
+      } catch (error) {
+        console.error('Error accepting private travels', error);
+      }
+      setAcceptedUser([]);
+    }
   }
 
   const cancelDelete = () => {
@@ -108,6 +134,7 @@ const PrivateTrip = () => {
 
   const deleteUser = (item) => {
     setSelectedUser(item);
+    setdeleteORaccept(false);
     setIsPopupVisible(true);
   };
 
@@ -120,12 +147,13 @@ const PrivateTrip = () => {
           <div className="links">
             <Link to="/travels/private/add">ADD</Link>
             <button onClick={handleDeleteSelected}>Delete Selected</button>
+            <button onClick={handleAcceptedSelected}>Accept Selected</button>
           </div>
         </div>
         <table>
           <thead>
             <tr>
-              <th>Goint Date</th>
+              <th>Going Date</th>
               <th>Going From</th>
               <th>Going To</th>
               <th>Bus ID</th>
@@ -189,7 +217,9 @@ const PrivateTrip = () => {
       </div>
       {isPopupVisible && (
         <Popup
-          message="Are you sure you want to delete the selected private travel?"
+          message={ deleteORaccept ? "Are you sure you want to accept the selected private travel?" : 
+            "Are you sure you want to delete the selected private travel?"
+          }
           onConfirm={confirmDelete}
           onCancel={cancelDelete}
         />
